Require a PcapNgState to encode or decode blocks.
Block::from_slice(), Block::write_to(), and all conversions from
RawBlock to Block require a PcapNgState, because without state
information the encoding is ambiguous.

This was previously handled in an ad-hoc manner in the reader, writer
and parser but is now centralised into the PcapNgState type.

The previous functions for adjusting EPB timestamps are removed as they
are no longer necessary. Timestamps are now converted as required in
from_slice() and write_to(), according to the current PcapNgState.